+++ modules/spring-core/src/main/kotlin/com/esprito/spring/core/inspections/SpringMetaAnnotationWithoutRuntimeInspection.kt
@@ -4,11 +4,13 @@ import com.esprito.inspection.SpringBaseUastLocalInspectionTool
 import com.esprito.spring.core.SpringCoreBundle
 import com.esprito.spring.core.SpringCoreClasses.RETENTION
 import com.esprito.spring.core.SpringCoreClasses.RETENTION_POLICY
+import com.esprito.spring.core.SpringCoreClasses.TARGET
 import com.esprito.spring.core.inspections.quickfix.RewriteAnnotationQuickFix
 import com.intellij.codeInsight.AnnotationUtil
 import com.intellij.codeInspection.InspectionManager
 import com.intellij.codeInspection.ProblemDescriptor
 import com.intellij.codeInspection.ProblemHighlightType
+import com.intellij.psi.PsiAnnotation
 import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiNameIdentifierOwner
@@ -22,17 +24,11 @@ class SpringMetaAnnotationWithoutRuntimeInspection : SpringBaseUastLocalInspecti
         manager: InspectionManager,
         isOnTheFly: Boolean
     ): Array<ProblemDescriptor>? {
-        //TODO: deep scan?
         if (!aClass.isAnnotationType) return null
         val psiClass = aClass.javaPsi
         val identity = getIdentifyingElement(aClass) ?: return null
 
-        if (psiClass
-                .annotations.none {
-                    it.qualifiedName
-                        ?.startsWith(SPRING_PREFIX) == true
-                }
-        ) return null
+        if (!isMetaAnnotatedWithSpringAnnotation(psiClass.annotations.toList())) return null
 
         val retention = psiClass.getAnnotation(RETENTION)
             ?: return createProblemDescriptor(identity, psiClass, manager, isOnTheFly)
@@ -46,6 +42,23 @@ class SpringMetaAnnotationWithoutRuntimeInspection : SpringBaseUastLocalInspecti
             createProblemDescriptor(identity, psiClass, manager, isOnTheFly)
         }
     }
+
+    private fun isMetaAnnotatedWithSpringAnnotation(psiAnnotations: Collection<PsiAnnotation>): Boolean {
+        for (psiAnnotation in psiAnnotations) {
+            val annotationQN = psiAnnotation.qualifiedName ?: continue
+            if (annotationQN in setOf(TARGET, RETENTION)) continue
+
+            if (annotationQN.startsWith(SPRING_PREFIX)) return true
+
+            val nestedAnnotations = psiAnnotation.resolveAnnotationType()
+                ?.annotations
+                ?.filter { it.qualifiedName != annotationQN }
+                ?: continue
+            if (isMetaAnnotatedWithSpringAnnotation(nestedAnnotations)) return true
+        }
+        return false
+    }
+
 
     private fun createProblemDescriptor(
         psiElement: PsiElement,