+++ CHANGELOG.md
@@ -7,6 +7,16 @@
 
 ### Spring Core/Boot
 
+- Added autocompletion in properties files for all classes fields
+- Inspection `unresolved property key` supports Relaxed Binding
+- Fixed inspection for `@Retention` on Spring annotations. Supported meta-annotations
+- Fixed stackoverflow error while property calculation on some cases
+- Support @Scheduled annotation
+
+## [241.1.1263] - 2024-04-17
+
+### Spring Core/Boot
+
 - Fixed inspection for constructor parameters with default value: Class constructor properties annotated
   with `@ConfigurationProperties` must be nullable
 - Property reference supports Relaxed Binding
@@ -17,7 +27,7 @@
 - Added license check scheduler
 - Fix: EDT exception in SpringToolRunConfigurationConfigurable
 
-## [241.1.1199] - 2024-04-11
+## 241.1.1199 - 2024-04-11
 
 ### Spring Core/Boot